Migrate establisher code to more relevant files
